+++ src/main/java/store/sonyk9919/api/global/config/RestClientConfig.java
@@ -1,21 +1,29 @@
 package store.sonyk9919.api.global.config;
 
+import java.net.http.HttpClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.http.client.JdkClientHttpRequestFactory;
+import org.springframework.web.client.RestClient;
 
 @Configuration
-public class WebClientConfig {
-
+public class RestClientConfig {
     @Value("${analysis.fastApi.url}")
     private String fastAPIUrl;
 
     @Bean
-    public WebClient fastAPIClient(){
-        return WebClient.builder()
+    public RestClient fastApiClient() {
+        HttpClient httpClient = HttpClient.newBuilder()
+                .version(HttpClient.Version.HTTP_1_1)
+                .build();
+
+        JdkClientHttpRequestFactory requestFactory = new JdkClientHttpRequestFactory(httpClient);
+
+        return RestClient.builder()
+                .requestFactory(requestFactory)
                 .baseUrl(fastAPIUrl)
                 .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
                 .build();
